Stop skipping SvROK handling for all magical scalars
Previously, for references with (non-tie, as references can't be tied) magic
attached, cloning of the referent has been skipped as magic_ref indicated that
it should be.

magic_ref appears to only have been intended to track whether or not a scalar is
tied so the contents of the tied structure aren't cloned. However, it actually
tracked whether or not a scalar had any kind of magic attached. That is wrong,
and this commit fixes it, making t/08fieldhash.t pass in the process.