removed dubious documentation on the optional parameter
The documentation says you can use this parameter to limit the depth
of the copy, but right now the extra parameter just increases the
refcount when it's 0, and clones otherwise. Until we decide which is
broken (the documentation or the implementation), we're removing the
mention to the option parameter. This fixes RT issue 57773.