Check the CowREFCNT of a COWed PV
Resolved #10

Strangely some PV could be cowed with
a COWRefcnt equal 0...

This might hide a more serious problem
somewhere else. But for now the best thing we
can do is leave these PVs alone.

Adjust the test as doing \"something" produces
a COWPV with CowREFCNT = 0 which is going
to be lost after cloning.