Add some COW test coverage
Provide a unit test which check cloning with COW
and reproduce the SEGV from GH #10.

This would allow us to refactor and build better COW
patch without hitting that regression a second time.

Note that the BUILD_REQUIRES was more a TEST_REQUIRES,
it's not really needed to use Test::More for building Clone.