license test files under GPL or Artistic
Unlike the rest of the distribution, which is licensed under the
same terms as Perl itself - meaning both Artistic License and GPL -
those 3 test files (which were originally borrowed from Storable's
test suite) were licensed solely under the Artistic License.
ppisar@redhat.com asked us to review this, since it meant they were
not able to distribute those files on Fedora. So we asked Raphael
Manfredi, the author of said files, and he graciously authorized
us to chance the wording so the files are explicitly released under
"the same terms as Perl itself".

This closes GH#20