Revert "Add a test to check without B::COW"
This partially reverts commit 7d9c699e962494fcdf7fd0c929c86195b29cb40f.

Optionally depending on B::COW is extra complexity, which was added
because it would not always be installed as a prereq. The issue with
B::COW not being seen as a prereq was due to the Makefile.PL not
including compatibilty code for older EUMM versions. This has been
fixed, so the module should always be available.