switch CI tests to install using cpanm to properly pick up prereqs
cpm can only install from a cpanfile, META file, or a fixed list of
prereqs. We want to install the prereqs from the Makefile.PL file, which
cpanm will do for us.

The install-with-cpanm sudo option is rather broken. It runs the cpanm
command itself with sudo, which means any files generated, like
MYMETA.json, are owned by root. This will prevent later commands from
working properly. cpanm's own sudo option is fine though, since it only
runs the install action with sudo.

At some point this should be changed prepare a dist dir or dist archive.
A dist dir or archive would contain a META file, allowing cpm to work.
It would make it easier to switch to a different author tool, and test on
a system or perl that can't run that author tool.

maybe less sudo?